+++ src/main/java/com/github/tier940/legacycraft/mixins/buildcraft/MixinAssemblyTableOutputFix.java
@@ -15,7 +15,7 @@ import buildcraft.lib.misc.InventoryUtil;
 import buildcraft.silicon.tile.TileAssemblyTable;
 
 @Mixin(value = TileAssemblyTable.class, remap = false)
-public class MixinTileAssemblyTable {
+public class MixinAssemblyTableOutputFix {
 
     @Redirect(
               method = "update",
@@ -23,10 +23,11 @@ public class MixinTileAssemblyTable {
                        value = "INVOKE",
                        target = "Lbuildcraft/lib/misc/InventoryUtil;addToBestAcceptor(Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/util/EnumFacing;Lnet/minecraft/item/ItemStack;)V",
                        remap = false))
-    private void lcCoreDeterministicOutput(World world, BlockPos pos, EnumFacing excludedFace, ItemStack stack) {
+    @SuppressWarnings("ConstantValue")
+    private void lc$deterministicOutput(World world, BlockPos pos, EnumFacing ignore, ItemStack stack) {
         for (EnumFacing side : EnumFacing.VALUES) {
             if (stack.isEmpty()) return;
-            if (side == excludedFace) continue;
+            if (side == ignore) continue;
             IInjectable injectable = ItemTransactorHelper.getInjectable(
                     world.getTileEntity(pos.offset(side)), side.getOpposite());
             if (injectable == null) continue;